+++ Trunk/src/scripts/src/GameMonkey/gm/gmGCRootUtil.h
@@ -121,13 +121,13 @@ public:
   // Return number of elements in array
   int GetSize()
   {
-    return m_array.size();
+    return (int)m_array.size();
   }
 
   // Return last index eg. GetSize()-1
   int GetLastIndex()
   {
-    return GetSize() - 1;
+    return int(GetSize() - 1);
   }
 
   // Remove element at index
